8051 Project
In this project, we implemented a program which continuously generates a sine wave (as an illustration) based on given inputs as amplitude and frequency. The pattern of a sine wave was introduced to the program as hard-coded array of length 128 including the periodic values of a sine wave. Even though this method does not lead to well precise drawings for smooth illustrations, it is the best solution considering the limited memory capacity of 8051 and limitation of the Keil uVision 5 platform. The sampling rate of the program is 1kHz independent from the given frequency meaning that the drawing will be updated 1000 times in a second which is sufficient enough for the frequency inputs less than 500 Hz according to Nyquist Sampling Theorem. The inputs can be taken from the user via the serial port or the keypad. The * symbol is used as an ending character for frequency input and # is used for signaling the ending of amplitude input. The accepted values are integers between the range (0, 999] which is then transformed into the range (0,255] proportionally. The sine wave drawn at the end represents the analog signal whose parameters are specified by the user as digital signal. The RRw/ISR architecture was used in the program design since the prioritization of the tasks is not significantly important for this problem.